+++ tmp/test_app/config/environments/production.rb
@@ -1,4 +1,7 @@
 TestApp::Application.configure do
+  # Precompile additional assets (application.js, application.css, and all non-JS/CSS are already added)
+  config.assets.precompile += %w(editing.css editing.js)
+
   # Settings specified here will take precedence over those in config/application.rb
 
   # Code is not reloaded between requests